+++ scripts/chat/services/index.js
@@ -21,4 +21,5 @@ export * from "./actions/hide-action.js";
 export * from "./actions/point-out-action.js";
 export * from "./actions/seek-action.js";
 export * from "./actions/sneak-action.js";
+export * from "./actions/take-cover-action.js";